Prepare project for public consumption
