+++ package.json
@@ -2,37 +2,60 @@
   "name": "billomat-ts",
   "version": "0.0.1",
   "description": "Billomat API implementation in TypeScript for use in Node.js",
-  "main": "src/index.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "dependencies": {
-    "@types/superagent": "^4.1.7",
-    "superagent": "^5.2.2",
-    "typescript": "^3.9.5"
+    "@types/superagent": "^4.1.10",
+    "superagent": "^6.1.0"
   },
   "devDependencies": {
-    "@types/chai": "^4.2.11",
-    "@types/mocha": "^7.0.2",
+    "@types/chai": "^4.2.14",
+    "@types/mocha": "^8.0.4",
     "@types/nock": "^11.1.0",
     "@types/proxyquire": "^1.3.28",
-    "@types/sinon": "^9.0.0",
+    "@types/sinon": "^9.0.8",
+    "@typescript-eslint/eslint-plugin": "^4.8.2",
+    "@typescript-eslint/eslint-plugin-tslint": "^4.8.2",
+    "@typescript-eslint/parser": "^4.8.2",
     "chai": "^4.2.0",
-    "mocha": "^7.2.0",
-    "nock": "^12.0.3",
+    "eslint": "^7.14.0",
+    "eslint-plugin-jsdoc": "^30.7.8",
+    "eslint-plugin-prefer-arrow": "^1.2.2",
+    "mocha": "^8.2.1",
+    "nock": "^13.0.5",
     "proxyquire": "^2.1.3",
-    "sinon": "^9.0.2",
-    "ts-node": "^8.10.1",
-    "tslint": "^6.1.2",
-    "tslint-eslint-rules": "^5.4.0"
+    "sinon": "^9.2.1",
+    "ts-node": "^9.0.0",
+    "typescript": "^4.1.2"
   },
   "scripts": {
-    "lint": "tslint -p .",
-    "pretest": "npm run lint",
-    "test": "mocha --require ts-node/register --extensions ts 'src/**/*.spec.ts'"
+    "lint": "eslint . --ext .js,.ts,.spec.js,.spec.ts",
+    "test": "mocha --require ts-node/register --extensions ts 'src/**/*.spec.ts'",
+    "build": "tsc",
+    "prepare": "npm run build",
+    "prepublishOnly": "npm test && npm run lint",
+    "preversion": "npm run lint",
+    "version": "git add -A src",
+    "postversion": "git push && git push --tags"
   },
   "author": "Seibert Media GmbH",
+  "keywords": [
+    "billomat",
+    "api",
+    "client",
+    "accounting"
+  ],
   "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/seibert-media/billomat-ts/issues"
+  },
+  "homepage": "https://github.com/seibert-media/billomat-ts/blob/master/README.md",
   "repository": {
     "type": "git",
     "url": "git@github.com:seibert-media/billomat-ts.git",
-    "directory": ""
-  }
+    "directory": "."
+  },
+  "files": [
+    "dist/**/*"
+  ]
 }
