--- src/billomat-client.d.ts
@@ -1,6 +1,0 @@
-export interface BillomatClientConfig {
-    baseUrl: string,
-    apiKey: string;
-    appSecret?: string,
-    appId?: string,
-}
