+++ src/index.ts
@@ -1,3 +1,3 @@
-import { billomatApi } from './billomat.api';
-
-export const billomatApiClient = billomatApi;
+export * from './get-billomat-api-client';
+export { Billomat } from './billomat';
+export { BillomatResourceClient } from './billomat-resource-client';
