+++ tsconfig.json
@@ -1,9 +1,20 @@
 {
   "compilerOptions": {
+    "declaration": true,
+    "module": "commonjs",
     "noImplicitReturns": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
+    "outDir": "./dist/",
     "strict": true,
-    "target": "ESNext"
-  }
+    "target": "es5"
+  },
+  "include": [
+    "src/**/*",
+    ".eslintrc.js"
+  ],
+  "exclude": [
+    "node_modules",
+    "dist/**/*"
+  ]
 }
