--- tslint.json
@@ -1,29 +1,0 @@
-{
-  "defaultSeverity": "error",
-  "extends": [
-    "tslint:recommended",
-    "tslint-eslint-rules"
-  ],
-  "jsRules": {},
-  "rules": {
-    "quotemark": [
-      true,
-      "single"
-    ],
-    "no-console": false,
-    "object-literal-shorthand": [
-      true,
-      "never"
-    ],
-    "object-literal-sort-keys": [
-      true,
-      "ignore-case"
-    ],
-    "typedef": [
-      true,
-      "call-signature",
-      "property-declaration"
-    ]
-  },
-  "rulesDirectory": []
-}
